Mahdi and Liquid

+++ drm/libdrmframework/include/PlugInManager.h
@@ -80,7 +80,7 @@ public:
         Vector<String8> plugInFileList = getPlugInPathList(rsPlugInDirPath);
 
         if (!plugInFileList.isEmpty()) {
-            for (size_t i = 0; i < plugInFileList.size(); ++i) {
+            for (unsigned int i = 0; i < plugInFileList.size(); ++i) {
                 loadPlugIn(plugInFileList[i]);
             }
         }
@@ -91,7 +91,7 @@ public:
      *
      */
     void unloadPlugIns() {
-        for (size_t i = 0; i < m_plugInIdList.size(); ++i) {
+        for (unsigned int i = 0; i < m_plugInIdList.size(); ++i) {
             unloadPlugIn(m_plugInIdList[i]);
         }
         m_plugInIdList.clear();
